Moves e2e tests for storage panel to officeUserPPM.js
