+++ cypress/integration/office/officeUserPPM.js
@@ -1,4 +1,6 @@
 /* global cy */
+import { userCancelsStorageDetails, userSavesStorageDetails } from '../../support/storagePanel';
+
 describe('office user finds the move', function() {
   beforeEach(() => {
     cy.signIntoOffice();
@@ -64,6 +66,16 @@ describe('office user finds the move', function() {
     officeUserGoesToPPMPanel('FDXTIU');
     officeUserEditsEstimatesPanel(60606, 72018);
   });
+
+  it('office user completes storage panel', function() {
+    officeUserGoesToStoragePanel('VGHEIS');
+    userSavesStorageDetails();
+  });
+
+  it('office user cancels storage panel', function() {
+    officeUserGoesToStoragePanel('NOADVC');
+    userCancelsStorageDetails();
+    })
 });
 
 function officeUserEditsNetWeight() {
@@ -406,4 +418,33 @@ function officeUserEditsEstimatesPanel(destinationPostalCode, pickupPostalCode) 
 
     cy.get('input[name="pickup_postal_code"]').should('have.value', `${pickupPostalCode}`);
   });
+}
+
+function officeUserGoesToStoragePanel(locator) {
+  // Open new moves queue
+  cy.location().should(loc => {
+    expect(loc.pathname).to.match(/^\/queues\/new/);
+  });
+
+  // Find shipment and open it
+  cy.selectQueueItemMoveLocator(locator);
+
+  cy.location().should(loc => {
+  expect(loc.pathname).to.match(/^\/queues\/new\/moves\/[^/]+\/basics/);
+});
+
+  cy
+  .get('.title')
+  .contains('PPM')
+  .click();
+
+  cy.location().should(loc => {
+  expect(loc.pathname).to.match(/^\/queues\/new\/moves\/[^/]+\/ppm/);
+});
+
+  cy
+  .get('.editable-panel-header')
+  .contains('Storage')
+  .siblings()
+  .click();
 }
